+++ src/backend/joanie/demo/management/commands/create_dev_demo.py
@@ -6,6 +6,7 @@ from django.utils import translation
 
 from joanie.core import enums, factories, models
 from joanie.core.models import CourseState
+from joanie.demo.defaults import NB_DEV_OBJECTS
 
 OPENEDX_COURSE_RUN_URI = (
     "http://openedx.test/courses/course-v1:edx+{course:s}+{course_run:s}/course"
@@ -60,23 +61,152 @@ class Command(BaseCommand):
                 )
         return courses
 
-    def _create_product(self, user, organization):
-        """Create a single product for given user and organization."""
-        course = self.create_course(user, organization)
-        target_courses = self.create_course(
-            user, organization, batch_size=3, with_course_runs=True
-        )
-        return factories.ProductFactory(
-            courses=[course],
-            target_courses=target_courses,
-            type=enums.PRODUCT_TYPE_CREDENTIAL,
-        )
-
-    def create_product(self, user, organization, batch_size=1):
+    def create_product_credential(self, user, organization, batch_size=1):
         """Create batch or products for given user and organization."""
         if batch_size == 1:
-            return self._create_product(user, organization)
-        return [self.create_product(user, organization) for i in range(batch_size + 1)]
+            course = factories.CourseFactory(
+                organizations=[organization],
+                users=[[user, enums.OWNER]],
+            )
+            product = factories.ProductFactory(
+                type=enums.PRODUCT_TYPE_CREDENTIAL,
+                courses=[course],
+            )
+            target_course_list = factories.CourseFactory.create_batch(
+                2,
+                organizations=[organization],
+                users=[[user, enums.OWNER]],
+            )
+
+            for target_course in target_course_list:
+                factories.ProductTargetCourseRelationFactory(
+                    course=target_course, product=product
+                )
+                factories.CourseRunFactory(
+                    course=target_course,
+                    is_listed=True,
+                    state=CourseState.ONGOING_OPEN,
+                    languages=self.get_random_languages(),
+                    resource_link=OPENEDX_COURSE_RUN_URI.format(
+                        course=target_course.code, course_run="{course.title}_run1"
+                    ),
+                )
+                factories.CourseRunFactory(
+                    course=target_course,
+                    is_listed=True,
+                    state=CourseState.ONGOING_OPEN,
+                    languages=self.get_random_languages(),
+                    resource_link=OPENEDX_COURSE_RUN_URI.format(
+                        course=target_course.code, course_run="{course.title}_run2"
+                    ),
+                )
+            self.stdout.write(
+                self.style.SUCCESS(
+                    f"Successfully create product credential on course {course.code}"
+                )
+            )
+            return product
+        return [
+            self.create_product_credential(user, organization)
+            for i in range(batch_size)
+        ]
+
+    def create_product_certificate(self, user, organization, batch_size=1):
+        """Create batch or products certificate for given user and organization."""
+        if batch_size == 1:
+            course = factories.CourseFactory(
+                organizations=[organization],
+                users=[[user, enums.OWNER]],
+            )
+            factories.CourseRunFactory(
+                course=course,
+                is_listed=True,
+                state=CourseState.ONGOING_OPEN,
+                languages=self.get_random_languages(),
+                resource_link=OPENEDX_COURSE_RUN_URI.format(
+                    course=course.code, course_run="{course.title}_run1"
+                ),
+            )
+            factories.CourseRunFactory(
+                course=course,
+                is_listed=True,
+                state=CourseState.ONGOING_OPEN,
+                languages=self.get_random_languages(),
+                resource_link=OPENEDX_COURSE_RUN_URI.format(
+                    course=course.code, course_run="{course.title}_run2"
+                ),
+            )
+            product = factories.ProductFactory(
+                type=enums.PRODUCT_TYPE_CERTIFICATE, courses=[course]
+            )
+            self.stdout.write(
+                self.style.SUCCESS(
+                    f"Successfully create product credential on course {course.code}"
+                )
+            )
+            return product
+        return [
+            self.create_product_certificate(user, organization)
+            for i in range(batch_size)
+        ]
+
+    def create_product_certificate_enrollment(self, user, organization):
+        """Create a product certificate and it's enrollment."""
+        product = self.create_product_certificate(user, organization)
+        course = product.courses.first()
+        return factories.EnrollmentFactory(
+            user=user,
+            course_run=course.course_runs.first(),
+            is_active=True,
+            state=enums.ENROLLMENT_STATE_SET,
+        )
+
+    def create_product_purchased(self, user, organization, product_type):
+        """Create a product certificate, it's enrollment and it's order."""
+        if product_type == enums.PRODUCT_TYPE_CERTIFICATE:
+            product = self.create_product_certificate(user, organization)
+        elif product_type == enums.PRODUCT_TYPE_CREDENTIAL:
+            product = self.create_product_credential(user, organization)
+        else:
+            raise ValueError(f"Given product_type ({product_type}) is not allowed.")
+
+        course = product.courses.first()
+
+        if product_type == enums.PRODUCT_TYPE_CERTIFICATE:
+            factories.EnrollmentFactory(
+                user=user,
+                course_run=course.course_runs.first(),
+                is_active=True,
+                state=enums.ENROLLMENT_STATE_SET,
+            )
+
+        return factories.OrderFactory(
+            state=enums.ORDER_STATE_VALIDATED,
+            product=product,
+            course=course,
+            owner=user,
+        )
+
+    def create_product_purchased_with_certificate(
+        self, user, organization, product_type
+    ):
+        """
+        Create a product, it's enrollment and it's order.
+        Also create the order's linked certificate.
+        """
+        order = self.create_product_purchased(user, organization, product_type)
+        return factories.OrderCertificateFactory(order=order)
+
+    def create_enrollment_certificate(self, user, organization):
+        """create an enrollment and it's linked certificate."""
+        course = self.create_course(user, organization, 1, True)
+        factories.EnrollmentCertificateFactory(
+            enrollment__user=user,
+            enrollment__course_run=course.course_runs.first(),
+            enrollment__is_active=True,
+            enrollment__state=enums.ENROLLMENT_STATE_SET,
+            organization=organization,
+        )
 
     def handle(self, *args, **options):  # pylint: disable=too-many-locals
         translation.activate("en-us")
@@ -159,20 +289,97 @@ class Command(BaseCommand):
 
         # Now create a course product to learn how to become a botanist and get a certificate
         # 1/ Create the credential Product linked to the botany Course
-        factories.ProductFactory(
+        product = factories.ProductFactory(
             type=enums.PRODUCT_TYPE_CREDENTIAL,
+            # organization=[organization],
             title="Become a certified botanist",
+            courses=[factories.CourseFactory(organizations=[organization])],
             target_courses=credential_courses,
             certificate_definition=factories.CertificateDefinitionFactory(
                 title="Botanist Certification",
                 name="Become a certified botanist certificate",
             ),
         )
+        self.stdout.write(
+            self.style.SUCCESS(f'Successfully create "{product.title}" product')
+        )
 
         # We need some pagination going on, let's create few more courses and products
         self.create_course(
-            admin_user, organization, batch_size=10, with_course_runs=True
+            admin_user,
+            organization,
+            batch_size=NB_DEV_OBJECTS["course"],
+            with_course_runs=True,
         )
-        self.create_product(admin_user, organization, batch_size=10)
+        self.stdout.write(
+            self.style.SUCCESS(
+                f"Successfully create {NB_DEV_OBJECTS['course']} fake courses"
+            )
+        )
+
+        self.create_product_credential(
+            admin_user, organization, batch_size=NB_DEV_OBJECTS["product_credential"]
+        )
+        self.stdout.write(
+            self.style.SUCCESS(
+                f"Successfully create {NB_DEV_OBJECTS['product_credential']} \
+                fake PRODUCT_CREDENTIAL"
+            )
+        )
+
+        self.create_product_certificate(
+            admin_user, organization, batch_size=NB_DEV_OBJECTS["product_certificate"]
+        )
+        self.stdout.write(
+            self.style.SUCCESS(
+                f"Successfully create {NB_DEV_OBJECTS['product_certificate']} \
+                fake PRODUCT_CERTIFICATE"
+            )
+        )
+
+        # Enrollments and orders
+        self.create_product_certificate_enrollment(admin_user, organization)
+        self.stdout.write(
+            self.style.SUCCESS(
+                "Successfully create a enrollment for a course with a PRODUCT_CERTIFICATE"
+            )
+        )
+
+        # Order for a PRODUCT_CERTIFICATE
+        self.create_product_purchased(
+            admin_user, organization, enums.PRODUCT_TYPE_CERTIFICATE
+        )
+        self.stdout.write(
+            self.style.SUCCESS("Successfully create a order for a PRODUCT_CERTIFICATE")
+        )
+
+        # Order for a PRODUCT_CERTIFICATE with a generated certificate
+        self.create_product_purchased_with_certificate(
+            admin_user, organization, enums.PRODUCT_TYPE_CERTIFICATE
+        )
+        self.stdout.write(
+            self.style.SUCCESS(
+                "Successfully create a order for a PRODUCT_CERTIFICATE \
+                with a generated certificate"
+            )
+        )
+
+        # Order for a PRODUCT_CREDENTIAL with a generated certificate
+        self.create_product_purchased_with_certificate(
+            admin_user, organization, enums.PRODUCT_TYPE_CREDENTIAL
+        )
+        self.stdout.write(
+            self.style.SUCCESS(
+                "Successfully create a order for a PRODUCT_CREDENTIAL with a generated certificate"
+            )
+        )
+
+        # Enrollment with a certificate
+        self.create_enrollment_certificate(admin_user, organization)
+        self.stdout.write(
+            self.style.SUCCESS(
+                "Successfully create a enrollment with a generated certificate"
+            )
+        )
 
         self.stdout.write(self.style.SUCCESS("Successfully fake data creation"))